Rename resolver
