--- tests/planners-impacts-test.csv
@@ -1,7 +1,0 @@
-"id";"condition";"category";"sector";"text1";"text2";"comment"
-"snow";"(temp_djf_iamean_s0p_hist <= -6)";;;"Internal rule";"Hamlet et al. (2007) broadly characterized snowmelt dominant basins as those with DJF below -6&deg;C, hybrid basins as those with DJF between -6&deg;C and 5&deg;C, and rain dominant basins those with a DJF exceeding 5&deg;C. Snow dominated basins are defined for this purpose as basins where the spatial median of temperature is less than or equal to -6&deg;C.";"internal snow rule"
-"hybrid";"((temp_djf_iamean_s0p_hist <= -6) && (temp_djf_iamean_s100p_hist >= -6)) || ((temp_djf_iamean_s0p_hist <= 5) && (temp_djf_iamean_s100p_hist >= 5)) || ((temp_djf_iamean_s0p_hist >= -6) && (temp_djf_iamean_s100p_hist <= 5))";;;"Internal rule";"Hybrid basins are basins which spatially vary sufficiently that they include temperature values between -6&deg;C and +5&deg;C.";"internal Hybrid rule"
-"rain";"(temp_djf_iamean_s100p_hist >= 5)";;;"Internal rule";"Rain-dominated basins are defined as basins where the spatial median temperature is greater than +5&deg;C; as per Hamlet et al.";"internal Rain rule"
-"future-snow";"(temp_djf_iamean_s0p_hist + temp_djf_iamean_s0p_e25p <= -6)";;;"Internal rule";"Internal rule";"internal future Snow rule"
-"future-hybrid";"((temp_djf_iamean_s0p_hist + temp_djf_iamean_s0p_e25p <= -6) && (temp_djf_iamean_s100p_hist + temp_djf_iamean_s100p_e75p >= -6)) || ((temp_djf_iamean_s0p_hist + temp_djf_iamean_s0p_e25p <= 5) && (temp_djf_iamean_s100p_hist + temp_djf_iamean_s100p_e75p >= 5)) || ((temp_djf_iamean_s0p_hist + temp_djf_iamean_s0p_e75p >= -6) && (temp_djf_iamean_s100p_hist + temp_djf_iamean_s100p_e25p <= 5))";;;"Internal rule";"Internal rule";"internal Hybrid rule"
-"future-rain";"(temp_djf_iamean_s100p_hist + temp_djf_iamean_s100p_e75p >= 5)";;;"Internal rule";"Internal rule";"internal Rain rule"
